requirements.txt
python-telegram-bot
motor==3.1.1
transformers==4.32.0
torch==2.2.0
opencv-python-headless==4.8.0
pymongo
pyrogram
uvloop
cryptg
tgcrypto
tensorflow
numpy
Pillow
opencv-python 
moviepy
lottie 
ffmpeg-python
aiofiles
dnspython